Factor gen-cmake subcommand into its own file
